added plan for refactor and made some small adjustments and added todos
